+++ libraries/transformer/src/main/java/androidx/media3/transformer/ExperimentalAnalyzerModeFactory.java
@@ -160,6 +160,11 @@ public final class ExperimentalAnalyzerModeFactory {
     }
 
     @Override
+    public Format getInputFormat() {
+      return configurationFormat;
+    }
+
+    @Override
     @Nullable
     public Format getOutputFormat() {
       return configurationFormat;